# docs/openapi.yaml
openapi: 3.1.0
info:
  title: vet402 API
  version: 1.0.0
  description: |
    Independent verification for the x402 agent-payment economy. Scores payee
    wallets and ERC-8004 agents on Base, returns a 0-100 score with an
    ALLOW/WARN/BLOCK recommendation, and publishes what happened after those
    verdicts.

    **Disclaimer:** Scores are informational only and do not constitute a guarantee,
    credit assessment, or investment advice. Every response that carries a score or
    a published rate repeats this string in the payload, so the caveat travels with
    the number.

    **Rate limits.** Authenticated calls: a per-account monthly quota (Free 1,000
    / Pro 50,000 / Scale 500,000), shared across keys, reset at the UTC month
    boundary. Each `/score` is 1 unit; a `/scores/batch` of N agents is N units.
    Scored responses carry `X-RateLimit-Limit`, `X-RateLimit-Used`, and
    `X-RateLimit-Remaining`. There is no per-second burst throttle on
    authenticated calls — you may spend the quota as fast as you like; a `429`
    with `Retry-After` means the monthly quota is exhausted. Separate IP-based
    abuse throttles (e.g. 60 auth-failures/min/IP) exist only to blunt
    key-guessing and scraping and are not hit by valid authenticated traffic.

    Key-less calls are throttled per IP instead. Every response on those paths
    carries `RateLimit-Limit` — the ceiling for that path, true for every caller.
...

...
    response.

  contact:
    name: vet402 API Support
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/wallets/{address}/score:
    get:
      operationId: getWalletScore
      tags: [Scores]
      summary: Get trust score by wallet address
      deprecated: true
      description: |
        Resolves wallet to agent when indexed; otherwise returns wallet heuristics.

        Deprecated since 2026-09-02: use
        `GET /api/v1/resources/{resourceId}/decision?role=payee&payer={address}`.
        The score's value is unchanged; since 2026-09-29 the response says so
        itself (`deprecated`, `deprecation`, and the `Deprecation` / `Link` headers).
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      responses:
        "200":
          description: Trust score result
          headers:
            Deprecation:
              $ref: "#/components/headers/ScoreDeprecation"
            Link:
              $ref: "#/components/headers/ScoreDeprecationLink"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ScoreDeprecationNotice"
                  - $ref: "#/components/schemas/TrustScoreResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/payees/{address}/score:
    get:
      operationId: getPayeeScore
      tags: [Payees]
      summary: Get trust score for a payee (payment-receiving wallet)
      deprecated: true
      description: |
        Deprecated since 2026-09-02: use
        `GET /api/v1/resources/{resourceId}/decision?role=payer` for the
        endpoint you are about to pay (`GET /api/v1/resolve?q={url}` finds its
        `resourceId`). This score rates the wallet as a whole, not that
        endpoint, and can say ALLOW where the decision says WARN or BLOCK.
        Its value is unchanged; since 2026-09-29 the response says so itself
        (`deprecated`, `deprecation`, and the `Deprecation` / `Link` headers).

        Buyer-side counterpart to `/api/v1/agents/{agentId}/score` and
        `/api/v1/wallets/{address}/score`,
        which screen the caller of an x402 route. This endpoint screens the *recipient*:
        "should my agent pay this wallet?" before an x402 payment is sent.

        Combines the wallet's own settlement receiving history (payments attested via
        `POST /api/v1/payments/x402` where it was the payee), wallet-age/tx-count health, an
        exit-scam-shaped outflow check (received funds then drained them), and any prior
        `verdict_outcomes` labels naming this wallet.

        Never 404s for an unfamiliar wallet — a wallet with little or no receiving
        history still returns `200` with `dataDepth: "thin"`, so integrators can decide
        how much to trust a thin-data score rather than being blocked outright.
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      responses:
        "200":
          description: Payee score result
          headers:
            Deprecation:
              $ref: "#/components/headers/ScoreDeprecation"
            Link:
              $ref: "#/components/headers/ScoreDeprecationLink"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ScoreDeprecationNotice"
                  - $ref: "#/components/schemas/PayeeScoreResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v1/payees/{address}/verdict-fast:
    get:
      operationId: getPayeeVerdictFast
      tags: [Scores]
      summary: verify-at-settle — cached verdict, never computes
      deprecated: true
      description: |
        Deprecated since 2026-09-02 with the payee score it caches; since
        2026-09-29 the response carries `deprecated`, `deprecation` and the
        `Deprecation` / `Link` headers.

        Returns the engine's cached, confident verdict or an honest
        `cache_cold` (treat as not-ALLOW; warm via the full /score endpoint).
        In-handler p95 is pinned under 1ms by the test suite. API key
        required (1 unit). See docs/verify-at-settle.md.
      responses:
        "200":
          description: A pinned verdict (`hit`) or an honest `cache_cold`
          headers:
            Deprecation:
              $ref: "#/components/headers/ScoreDeprecation"
            Link:
              $ref: "#/components/headers/ScoreDeprecationLink"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ScoreDeprecationNotice"
                  - $ref: "#/components/schemas/PayeeVerdictFast"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { description: Verdict surface unavailable }
...
components:
  headers:
    ScoreDeprecation:
      description: RFC 9745 deprecation date of the 0–100 score (2026-09-02T00:00:00Z).
      schema: { type: string, example: "@1788307200" }
    ScoreDeprecationLink:
      description: Where the deprecation is explained (`rel="deprecation"`).
      schema: { type: string, example: '<https://vet402.com/docs/api#endpoints>; rel="deprecation"; type="text/html"' }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key issued from the developer dashboard
  parameters:
...
    WalletAddress:
      name: address
      in: path
      required: true
      description: EVM wallet address (checksummed or lowercase)
      schema:
        type: string
        pattern: "^0x[a-fA-F0-9]{40}$"
        example: "0x1234567890123456789012345678901234567890"
...
  schemas:
    ScoreDeprecationNotice:
      type: object
      description: |
        Carried by every 0–100 score response since 2026-09-29. The score
        itself is unchanged; this says it is deprecated and what to call instead.
      required: [deprecated, deprecation]
      properties:
        deprecated:
          type: boolean
          enum: [true]
        deprecation:
          type: object
          required: [since, use_instead, find_resource_id, message]
          properties:
            since: { type: string, format: date, example: "2026-09-02" }
            use_instead: { type: string }
            find_resource_id: { type: string }
            message: { type: string }
    Recommendation:
      type: string
      enum: [ALLOW, WARN, BLOCK]
    TrustSignals:
      type: object
      required: [identity, reputation, wallet, x402, sybil, manual]
      properties:
        identity:
          type: object
          properties:
            registered: { type: boolean }
            hasMetadataUri:
              type: boolean
              description: Whether the agent has a metadata URI on-chain (not a validation of URI contents)
        reputation:
          type: object
          properties:
            feedbackCount: { type: integer }
            avgScore: { type: number }
            onChainAvgScore:
              type: number
              description: Raw on-chain reputation average before normalization
        wallet:
          type: object
          properties:
            ageDays: { type: integer }
            txCount: { type: integer }
            isBurner: { type: boolean }
        x402:
          type: object
          properties:
            paymentCount: { type: integer }
            uniqueDays: { type: integer }
            score:
              type: number
              description: >-
                Internal 0–100 component for the highest-weighted axis.
                vet402 2026-08-14: the axis is "verifiable economic activity",
                not x402 settlements alone — this reflects the STRONGER of the
                L1 observed purchases below and the x402 settlements above.
                Kept under `x402` for wire back-compat.
            l1PurchaseCount:
              type: integer
              description: >-
                vet402 2026-08-14 — delivery-verified L1 observed purchases
                behind the axis score (the premium signal). 0 until the
                observatory writes its first row for this wallet.
            l1DistinctSellers:
              type: integer
              description: Distinct L1 counterparties behind l1PurchaseCount.
        sybil:
          type: object
          properties:
            risk:
              type: string
              enum: [low, medium, high]
            flags:
              type: array
              items: { type: string }
        manual:
          type: object
          properties:
            list:
              type: string
              enum: [none, whitelist, blacklist]
    TrustScoreResult:
      type: object
      required:
        - agentId
        - trustScore
        - recommendation
        - signals
        - scoredAt
        - cacheExpiresAt
        - disclaimer
      properties:
        agentId:
          type: string
          description: ERC-8004 agent ID; "0" if wallet-only lookup
        wallet:
          type: string
          nullable: true
        trustScore:
          type: integer
          minimum: 0
          maximum: 100
        recommendation:
          $ref: "#/components/schemas/Recommendation"
        signals:
          $ref: "#/components/schemas/TrustSignals"
        breakdown:
          $ref: "#/components/schemas/ScoreBreakdown"
        reasons:
          type: array
          items: { type: string }
          description: >-
            Stable machine-readable codes behind the verdict, derived from the
            same signals (for integrators' compliance logs). Cannot disagree
            with the recommendation.
        scoredAt:
          type: string
          format: date-time
        cacheExpiresAt:
          type: string
          format: date-time
        disclaimer:
          type: string
        manualOverride:
          type: boolean
          description: True when customer WL/BL changed the recommendation
        blockReason:
          type: string
          description: Present when recommendation is BLOCK (e.g. wallet_mismatch, operator_policy)
        dataCoverage:
          type: object
          description: Indexer and settlement data freshness
          properties:
            ownerIndexer:
              type: object
              properties:
                status:
                  type: string
                  enum: [synced, partial, unavailable]
                blocksBehind:
                  type: integer
                  nullable: true
                lastBlock:
                  type: string
                  nullable: true
                  description: Last block the owner indexer has processed (decimal string); null when unknown
                indexedAgentRows:
                  type: integer
                  description: Agent rows the owner indexer currently holds
                staleRisk: { type: boolean }
            settlement:
              type: object
              properties:
                paymentRows: { type: integer }
                distinctWallets: { type: integer }
                recentPayments30d: { type: integer }
                walletHasHistory: { type: boolean }
...
    ScoreBreakdown:
      type: object
      description: >-
        N-21 explainability. Decomposition of the chain score into its four
        weighted components plus the sybil adjustment — the exact arithmetic of
        the weighting, so it cannot disagree with trustScore. Omitted on
        hard-blocked verdicts where no weighting ran. The manual whitelist/
        blacklist layer is intentionally excluded (see manualOverride).
      required: [components, weightedSubtotal, sybilPenalty, prePolicyScore]
      properties:
        components:
          type: object
          required: [identity, reputation, wallet, x402]
          properties:
            identity: { $ref: "#/components/schemas/ScoreComponent" }
            reputation: { $ref: "#/components/schemas/ScoreComponent" }
            wallet: { $ref: "#/components/schemas/ScoreComponent" }
            x402: { $ref: "#/components/schemas/ScoreComponent" }
        weightedSubtotal:
          type: number
          description: Weighted average of the four components before the sybil penalty
        sybilPenalty:
          type: number
          description: Points removed by sybil/availability flags (<= 0); see signals.sybil.flags
        prePolicyScore:
          type: integer
          minimum: 0
          maximum: 100
          description: weightedSubtotal + sybilPenalty, clamped; equals trustScore unless a manual list moved it
...
    DataDepth:
      type: string
      enum: [thin, moderate, rich]
      description: |
        How much receiving history this score could draw on. `thin` wallets lean more
        heavily on wallet-health and outflow-pattern signals since there isn't much
        settlement track record yet; `rich` wallets are judged mostly on that track
        record. Integrators should treat a `thin` score as lower-confidence.
    PayeeSignals:
      type: object
      required: [receiving, walletHealth, drainPattern, outcomeHistory, flags]
      properties:
        receiving:
          type: object
          description: Settlement history where this wallet was the payee (attested via POST /payments/x402)
          properties:
            paymentCount: { type: integer }
            uniqueDays: { type: integer }
            distinctPayers: { type: integer }
            score:
              type: number
              description: Internal 0-100 receiving-track-record component
            l1DeliveryCount:
              type: integer
              description: >-
                vet402 2026-08-14 — delivery-verified L1 receipts behind the
                receiving score (the premium signal: the observatory actually
                paid this payee and checked what came back). 0 until the
                observatory writes its first row for this wallet.
            l1DistinctBuyers:
              type: integer
              description: Distinct L1 buyers behind l1DeliveryCount.
            l1Settled:
              type: integer
              description: >-
                Settlements to this payee that vet402 confirmed on-chain
                (recipient, amount, token, chain, confirmations).
            l1PaidNeverSettled:
              type: integer
              description: >-
                Paid attempts by vet402 that resolved WITHOUT an on-chain
                settlement. Pending verifications are not counted here.
            l1NonSettlingDays:
              type: integer
              description: >-
                Distinct UTC days on which a non-settling paid attempt
                occurred. Separates a one-day outage from a sustained
                pattern.
            l1PendingVerification:
              type: integer
              description: >-
                Paid attempts whose claimed settlement vet402 has not
                re-read on-chain yet. Never counted against the payee.
            l1NonDeliveryReason:
              type: string
              nullable: true
              enum: [paid_never_settled_sustained, paid_never_settled_repeated, paid_never_settled]
              description: >-
                Present when a non-delivery ceiling was applied to the final
                score. null when no such ceiling applies.
        walletHealth:
          type: object
          description: Reused from the same on-chain wallet-age/tx-count/burner heuristics as the seller-side score
          properties:
            ageDays: { type: integer }
            txCount: { type: integer }
            isBurner: { type: boolean }
            score:
              type: number
              description: Internal 0-100 wallet-health component
        drainPattern:
          type: object
          description: Exit-scam-shaped check — received funds, then pulled out (near-)everything
          properties:
            detected: { type: boolean }
            drainRatio:
              type: number
              nullable: true
              description: outgoing / (outgoing + current balance) over the wallet's recent transactions
            outgoingCount: { type: integer }
            incomingCount: { type: integer }
            score:
              type: number
              description: Internal 0-100 drain-pattern component
            unmeasured:
              type: array
              items: { type: string }
              description: >-
                Asset legs that could NOT be read on this request, named rather
                than averaged away — e.g. ["native_drain"], ["usdc_drain"], or
                both. Empty when every leg was assessed. These names also
                appear in the top-level `signalsUnavailable`; a non-empty list
                means the drain view is partial.
        outcomeHistory:
          type: object
          description: Prior verdict_outcomes labels naming this wallet (auto-detected or partner-reported)
          properties:
            types:
              type: array
              items: { type: string }
            adjustment:
              type: number
              description: Net point adjustment applied to the pre-outcome weighted score
        flags:
          type: array
          items: { type: string }
          description: e.g. wallet_metrics_unavailable, drain_check_unavailable, new_burner_wallet, negative_outcome:rug_pull_outflow
    PayeeVerdictFast:
      oneOf:
        - $ref: "#/components/schemas/PayeeVerdictFastHit"
        - $ref: "#/components/schemas/PayeeVerdictFastCold"
      discriminator:
        propertyName: status
      description: |
        verify-at-settle answer. **Only `hit` + `ALLOW` clears a payment.**
        `cache_cold` is the ABSENCE of a verdict, not a permissive one — treat
        it as not-ALLOW, warm asynchronously via the full `/score` endpoint
        (same cache, TTL 5 min), and retry. See docs/verify-at-settle.md; the
        rule is implemented as `payeeVerdictFastAllows` in @vet402/sdk.

        This body deliberately carries no `degraded` / `signalsUnavailable`:
        the engine never pins a degraded or partially-measured reading, so a
        `hit` cannot be a fail-closed refusal wearing an ALLOW.
    PayeeVerdictFastHit:
      type: object
      required: [status, recommendation, score, cacheExpiresAt, handlerMicros]
      properties:
        status:
          type: string
          enum: [hit]
        recommendation:
          $ref: "#/components/schemas/Recommendation"
        score: { type: integer, minimum: 0, maximum: 100 }
        cacheExpiresAt:
          type: string
          format: date-time
          description: >-
            When this pinned verdict stops being the current answer. Clients
            gating money should re-check it rather than assume the server
            filtered expired entries (it does — this is belt and braces).
        handlerMicros:
          type: number
          description: In-handler microseconds spent (a cache read plus JSON)
    PayeeVerdictFastCold:
      type: object
      required: [status, recommendation, warmVia, note, handlerMicros]
      properties:
        status:
          type: string
          enum: [cache_cold]
        recommendation:
          type: string
          nullable: true
          enum: [null]
          description: Always null — nothing was pinned. NOT an allow.
        warmVia:
          type: string
          description: Full-score path to call fire-and-forget to warm the same cache
        note: { type: string }
        handlerMicros: { type: number }
...
    PayeeScoreResult:
      type: object
      required:
        - payee
        - score
        - recommendation
        - dataDepth
        - degraded
        - signalsUnavailable
        - signals
        - scoredAt
        - cacheExpiresAt
        - disclaimer
      properties:
        payee:
          type: string
          description: Lowercased wallet address that was scored
        score:
          type: integer
          minimum: 0
          maximum: 100
        recommendation:
          $ref: "#/components/schemas/Recommendation"
        dataDepth:
          $ref: "#/components/schemas/DataDepth"
        degraded:
          type: boolean
          description: >-
            **Read this before you act on `recommendation`.** True when at
            least one input could not be read at all, so this body is a
            fail-closed refusal, not a measurement. `dataDepth` answers "how
            much history does this wallet have?"; this answers "did we manage
            to look?".

            A client MUST NOT treat a `degraded: true` body as ALLOW, whatever
            `recommendation` and `score` say. The engine already forces
            BLOCK here and refuses to cache the verdict; @vet402/sdk
            (SpendGuard, `payee_score_degraded`) and @vet402/middleware
            (`score_degraded`) both deny on this field alone, and so does the
            Python client in the repository (`packages/python-sdk`, not
            published to PyPI). The `@vouchscore/*` packages are the former,
            deprecated names of the same npm packages. Always present.
        signalsUnavailable:
          type: array
          items: { type: string }
          description: >-
            **Read this before you act on `recommendation`.** Every input that
            could not be read on this request, named — currently one or more of
            `wallet_metrics`, `native_drain`, `usdc_drain`, `outcome_history`.
            Empty means the whole assessment was measured. Treat the list as
            open: new input names may be added without a breaking change.

            Non-empty with `degraded: false` is the partial-measurement case:
            the score IS backed by real measurements, just not all of them, and
            it has been capped below ALLOW for exactly that reason. A client
            that must not act on an incomplete view MUST NOT treat a body with
            a non-empty `signalsUnavailable` as ALLOW — SpendGuard denies it as
            `payee_partial_measurement` and the middleware as
            `partial_measurement`. Always present (empty array when clean).
        signals:
          $ref: "#/components/schemas/PayeeSignals"
        scoredAt:
          type: string
          format: date-time
        cacheExpiresAt:
          type: string
          format: date-time
        disclaimer:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: |
            Stable machine-readable error code. Request/body validation failures
            (malformed JSON, failed schema checks, empty/oversized batches, etc.)
            always use `invalid_request`.
          enum:
            - feed_unavailable
            - no_stock_token_activity
            - too_busy
            - facilitator_unavailable
            - too_slow_retry
            - invalid_request
            - subscribe_unavailable
            - db_unavailable
            - honeypot
            - invalid_email
            - invalid_endpoint
            - invalid_kind
            - reason_length
            - reason_required
            - invalid_agent_id
            - invalid_wallet_address
            - invalid_tx_hash
            - attestation_unverifiable
            - missing_api_key
            - invalid_api_key
            - auth_unavailable
            - rate_limit_exceeded
            - plan_upgrade_required
            - scoring_unavailable
            - payment_ingest_unavailable
            # Key-less verification, watchlist, accuracy and the JSON 404
            # (2026-08-13: these paths were live but undocumented, so their
            # codes were missing from an enum a generated client validates on).
            - invalid_name
            - invalid_id
            - url_must_be_https
            - agent_wallet_unbound
            - signature_mismatch
            - chain_unavailable
            - store_unavailable
            - accuracy_unavailable
            - watchlist_unavailable
            - limit_reached
            - rate_limited
            - not_found
            - forbidden
            # 2026-09-02 敵対的監査 P1-1: 実装が返しているのに enum に無かった語を全数
            # 追加。tests/openapi-error-enum.test.ts が「文書化された route.ts の
            # error リテラル ⊆ この enum」を機械検査する。
            # --- §7.3 / §9.1 resolve / resources / endpoints / payees / facts ---
            - invalid_query
            - resolve_unavailable
            - invalid_resource_id
            - invalid_endpoint_id
            - invalid_payee_id
            - unavailable
            # --- §9.1 /decision ---
            - invalid_role
            - invalid_caller_dialect
            - payer_required
            - decision_unavailable
            # caller policy queries (2026-09-07): the SDK's own caller-error words
            - invalid_amount_usd
            - invalid_policy
            - invalid_evidence_policy
            # --- §9.1 /census/summary ---
            - invalid_window
            - invalid_chain
            # --- scoring / verify / outcome ---
            - unsupported_chain
            - signature_expired
            - stale_signature
            - invalid_trust_event_id
            - outcome_ingest_unavailable
            - trust_event_not_found
            - related_wallet_mismatch
            - self_affirmation_not_accepted
            # --- webhooks ---
            - webhooks_unavailable
            - invalid_webhook_url
            - unknown_event
            # --- demo / contributions / requests / disputes / waitlist (JSON bodies) ---
            - invalid_json
            - invalid_input
            - endpoint_id_required
            - invalid_level
            - demo_l1_disabled
            - demo_budget_exhausted
            # 2026-09-04: デモ経由の L1 実購入が cron と同じ l1-purchase リースを取るようになった（409）。
            - l1_busy
            # 2026-09-05: 実行時の支出停止スイッチが立っている（503）。障害ではなく運用者の停止。
            - spending_halted
            - demo_unavailable
            - contributions_unavailable
            - requests_unavailable
            - disputes_unavailable
            - waitlist_unavailable
            # --- observatory reads ---
            - invalid_days
            - observatory_unavailable
            - endpoint_not_found
            - frame_unavailable
            - graph_unavailable
            - flags_unavailable
            - verdict_unavailable
            # --- observatory/watch (signed subscriptions) ---
            - invalid_signature
            - invalid_token
            - service_unavailable
            - watcher_limit_reached
            - internal_error
            # 2026-09-05 (S-6): 署名本文が vet402.com を名乗り domain 行を持つ形へ
            # 変わった。2026-09-05 以前の本文への署名は 2026-09-21T00:00:00Z まで
            # 受理し、それ以降はこのコードで拒否する——無言の signature_mismatch に
            # すると、正しく署名できている相手が理由を知る手段を失う。
            - signature_message_legacy_expired
            # --- badge ---
            - invalid_address
        details:
          type: object
          description: Present for `invalid_request` — Zod validation error detail (flattened).
        reason:
          type: string
          description: Present for `attestation_unverifiable` — human-readable reason the on-chain check failed.
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: |
        Rate limit exceeded. On authenticated paths this means the monthly quota
        is spent (`rate_limit_exceeded`); on key-less paths it is the per-IP
        window (`rate_limited`).
      headers:
        Retry-After:
          description: Seconds until the window resets.
          schema: { type: integer }
        RateLimit-Limit:
          description: Key-less paths — ceiling for the IP window. Sent on every response.
          schema: { type: integer }
        RateLimit-Remaining:
          description: >-
            Key-less paths — requests left in your current IP window. Per-caller,
            so it is sent here and on other uncached responses, but not on a
            CDN-cached 200.
          schema: { type: integer }
        RateLimit-Reset:
          description: >-
            Key-less paths — unix epoch seconds when your window resets. Same
            per-caller rule as RateLimit-Remaining.
          schema: { type: integer }
        X-RateLimit-Limit:
          description: Authenticated paths — monthly plan quota.
          schema: { type: integer }
        X-RateLimit-Used:
          description: Authenticated paths — quota consumed this month.
          schema: { type: integer }
        X-RateLimit-Remaining:
          description: Authenticated paths — quota left this month.
          schema: { type: integer }
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: rate_limit_exceeded
              retryAfter:
                type: integer